requirements.txt
asciitree
bcrypt
colorama
prompt_toolkit
pycryptodomex
pyperclip
tabulate
websockets
fido2>=2.0.0; python_version>='3.10'
requests>=2.31.0
cryptography>=39.0.1
protobuf>=4.23.0
keeper-secrets-manager-core>=16.6.0
keeper_pam_webrtc_rs; python_version>='3.8'
pydantic>=2.6.4; python_version>='3.8'
flask; python_version>='3.8'
pyngrok; python_version>='3.8'
flask-limiter; python_version>='3.8'
psutil; python_version>='3.8'
python-dotenv
fpdf2>=2.8.3
cbor2; sys_platform == "darwin" and python_version>='3.10'
pyobjc-framework-LocalAuthentication; sys_platform == "darwin" and python_version>='3.10'
winrt-runtime; sys_platform == "win32"
winrt-Windows.Foundation; sys_platform == "win32"
winrt-Windows.Security.Credentials.UI; sys_platform == "win32"